Finnished Assignment 1.06
Added Fog of war, an 'f' command that toggles it, and a 'g' command that
enables teleportation. Last monster locations are not seen, the fog of
war does not appear while choosing where to teleport to. As per the
spec, the mosnter list is completely unchanged, and thus lists all monsters.